Update README.md
Space Station Site 

This is an example home page for a Space travel company. Its styled using CSS and contains a navbar, imagine containers, and a form.

Shirt Station Site This is an example home page for a Shirt selling company. Its styled using CSS and contains a navbar, image containers, and a form.

Movie Theater Site

This page is an example for a Movie Theater that is styled using Boostrap 5. It contains many Boostrap components including but not limited to a navbar, jumbotron, form, cards, and dropdowns.

Portfolio Site

This is an example home page for my Business Portfolio. Styled using CSS and HTML which contains a navbar, image containers, and a form.